omp: fix number of threads allocation and add initialize from the file
It seems that by OpenMP standard omp_set_dynamic(TRUE) let the system decide
how many threads to allocate at the beginning of each parallel section and thus
the number of threads is changing.  We want to have a control over the number
so set to FALSE.
NOTE: ICC compiler has the opposite behavior which I didnt take into account.